feat: run app seed only in dev environment

+++ src/seed/seed.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PokeResponse } from './interfaces/poke-response.interface';
 import { PokemonService } from 'src/pokemon/pokemon.service';
 import { CreatePokemonDto } from 'src/pokemon/dtos';
@@ -9,6 +9,7 @@ import { ConfigService } from '@nestjs/config';
 export class SeedService {
   private readonly _defaultLimit: number;
   private readonly _baseUrl: string;
+  private readonly _environment: string;
 
   constructor(
     private readonly pokemonService: PokemonService,
@@ -17,9 +18,15 @@ export class SeedService {
   ) {
     this._defaultLimit = this.configService.get<number>('defaultSeedLimit')!;
     this._baseUrl = this.configService.get<string>('apiBaseUrl')!;
+    this._environment = this.configService.get<string>('environment')!;
   }
 
   async loadData() {
+    if (this._environment !== 'dev')
+      throw new BadRequestException(
+        `This option is only available for dev environment`,
+      );
+
     await this.pokemonService.deleteAllAsync();
 
     const data = await this.axiosAdapter.get<PokeResponse>(
